requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
pydantic==2.4.2
beanie==1.22.6
motor==3.3.1
python-dotenv==1.0.0
python-multipart==0.0.6 